+++ app/api/file.js
@@ -1,9 +1,15 @@
 "use strict";
 
 const { fileEmbed } = require("../../shared/mimeClass");
+const { optionalQuery } = require("../utils/optionalQuery");
 
 function canvasPath(request) {
-  return `/api/v1/files/${request.params.fileId}?include[]=preview_url`;
+  return `/api/v1/files/${
+    request.params.fileId
+  }?include[]=preview_url${optionalQuery(
+    request.query,
+    "replacement_chain_context_type"
+  )}${optionalQuery(request.query, "replacement_chain_context_id")}`;
 }
 
 function canvasResponseHandler(request, response, canvasResponse) {
@@ -16,7 +22,7 @@ function canvasResponseHandler(request, response, canvasResponse) {
       name: file.display_name || file.filename,
       url: file.url,
       preview_url: file.preview_url,
-      embed: fileEmbed(file)
+      embed: fileEmbed(file),
     });
   } else {
     response.send(canvasResponse.body);