+++ test/service/api/file.test.js
@@ -6,12 +6,25 @@ const { canvasPath, canvasResponseHandler } = require("../../../app/api/file");
 
 describe("File API", () => {
   describe("canvasPath()", () => {
+    const id = 47;
+    const params = { fileId: id };
+
     it("builds the correct path including file id", () => {
-      const id = 47;
-      const params = { fileId: id };
       const query = { per_page: 50 };
       const expectedPath = `/api/v1/files/${id}?include[]=preview_url`;
       assert.equal(canvasPath({ params, query }), expectedPath);
+    });
+
+    describe("when replacement context params are given", () => {
+      const query = {
+        replacement_chain_context_type: "course",
+        replacement_chain_context_id: 2,
+      };
+
+      it("includes the replacement context params in the query string", () => {
+        const expectedPath = `/api/v1/files/47?include[]=preview_url&replacement_chain_context_type=course&replacement_chain_context_id=2`;
+        assert.equal(canvasPath({ params, query }), expectedPath);
+      });
     });
   });
 
@@ -22,11 +35,11 @@ describe("File API", () => {
       request = {};
       response = {
         status: sinon.spy(),
-        send: sinon.spy()
+        send: sinon.spy(),
       };
       canvasResponse = {
         status: 200,
-        body: []
+        body: [],
       };
     });
 
@@ -51,14 +64,14 @@ describe("File API", () => {
           display_name: "Foo",
           filename: "Foo.pdf",
           preview_url: "someCANVADOCSurl",
-          url: "someurl"
+          url: "someurl",
         };
       });
 
       it("file has correctly tranformed properties", () => {
         canvasResponse.body = [file];
         canvasResponseHandler(request, response, canvasResponse);
-        response.send.calledWithMatch(val => {
+        response.send.calledWithMatch((val) => {
           return sinon.match(
             {
               id: file.id,
@@ -66,7 +79,7 @@ describe("File API", () => {
               name: file.display_name,
               url: file.url,
               preview_url: file.preview_url,
-              embed: { type: "file" }
+              embed: { type: "file" },
             },
             val
           );
@@ -77,7 +90,7 @@ describe("File API", () => {
         file.display_name = undefined;
         canvasResponse.body = [file];
         canvasResponseHandler(request, response, canvasResponse);
-        response.send.calledWithMatch(val => {
+        response.send.calledWithMatch((val) => {
           return sinon.match({ name: file.filename }, val);
         });
       });